fix: make repository action CI portable

+++ crates/jig/tests/vault_tui.rs
@@ -390,7 +390,7 @@ fn make_stdin_controlling_terminal(command: &mut Command) {
             if libc::setsid() == -1 {
                 return Err(std::io::Error::last_os_error());
             }
-            if libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY, 0) == -1 {
+            if libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY as libc::c_ulong, 0) == -1 {
                 return Err(std::io::Error::last_os_error());
             }
             Ok(())
